requirements.txt
numpy
pandas
streamlit
google-cloud-bigquery
google-generativeai
db-dtypes
python-dotenv
plotly
